Refactor native directories
