+++ SIT.Manager/Native/Linux/Managers/DxvkManager.cs
@@ -1,6 +1,7 @@
+using SIT.Manager.Native.Linux;
 using System.Collections.Generic;
 
-namespace SIT.Manager.Linux.Managers;
+namespace SIT.Manager.Native.Linux.Managers;
 
 public class DxvkManager() : DllManager("DXVK",
     [
