+++ SIT.Manager/Native/Linux/Managers/Vkd3DManager.cs
@@ -1,6 +1,7 @@
+using SIT.Manager.Native.Linux;
 using System.Collections.Generic;
 
-namespace SIT.Manager.Linux.Managers;
+namespace SIT.Manager.Native.Linux.Managers;
 
 public class Vkd3DManager() : DllManager("VKD3D",
     [
