+++ SIT.Manager/Native/Windows/WindowsApi.cs
@@ -1,37 +1,37 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 
-namespace SIT.Manager.Native;
+namespace SIT.Manager.Native.Windows;
 
 public static class WindowsApi
 {
     [DllImport("user32.dll", SetLastError = true)]
-    public static extern IntPtr SetParent(IntPtr hWndChild, IntPtr hWndNewParent);
+    public static extern nint SetParent(nint hWndChild, nint hWndNewParent);
 
     [DllImport("user32.dll", SetLastError = true)]
-    public static extern uint GetWindowLongPtr(IntPtr hWnd, int nIndex);
+    public static extern uint GetWindowLongPtr(nint hWnd, int nIndex);
 
     [DllImport("user32.dll", SetLastError = true, EntryPoint = "SetWindowLong")]
-    private static extern uint SetWindowLong32b(IntPtr hWnd, int nIndex, uint value);
+    private static extern uint SetWindowLong32b(nint hWnd, int nIndex, uint value);
 
     [DllImport("user32.dll", SetLastError = true, EntryPoint = "SetWindowLongPtr")]
-    private static extern IntPtr SetWindowLong64b(IntPtr hWnd, int nIndex, IntPtr value);
+    private static extern nint SetWindowLong64b(nint hWnd, int nIndex, nint value);
 
     [DllImport("user32.dll", EntryPoint = "SetWindowLongPtr")]
-    private static extern IntPtr SetWindowLongPtr64(HandleRef hWnd, int nIndex, IntPtr dwNewLong);
+    private static extern nint SetWindowLongPtr64(HandleRef hWnd, int nIndex, nint dwNewLong);
 
     [DllImport("user32.dll", EntryPoint = "SetWindowLong")]
     private static extern int SetWindowLong32(HandleRef hWnd, int nIndex, int dwNewLong);
 
-    public static IntPtr SetWindowLongPtr(HandleRef hWnd, int nIndex, IntPtr dwNewLong)
+    public static nint SetWindowLongPtr(HandleRef hWnd, int nIndex, nint dwNewLong)
     {
-        if (IntPtr.Size == 8)
+        if (nint.Size == 8)
         {
             return SetWindowLongPtr64(hWnd, nIndex, dwNewLong);
         }
         else
         {
-            return new IntPtr(SetWindowLong32(hWnd, nIndex, dwNewLong.ToInt32()));
+            return new nint(SetWindowLong32(hWnd, nIndex, dwNewLong.ToInt32()));
         }
     }
 }
