+++ SIT.Manager/Services/TarkovClientService.cs
@@ -1,8 +1,8 @@
 ﻿using Avalonia.Controls.ApplicationLifetimes;
 using SIT.Manager.Interfaces;
-using SIT.Manager.Linux;
 using SIT.Manager.ManagedProcess;
 using SIT.Manager.Models;
+using SIT.Manager.Native.Linux;
 using System;
 using System.Diagnostics;
 using System.IO;
